Moving plotting into the utils folder (#719)
